+++ apps/api/src/health/indicators/redis.health.ts
@@ -1,7 +1,7 @@
 import { CACHE_MANAGER } from '@nestjs/cache-manager'
 import { Inject, Injectable } from '@nestjs/common'
 import { HealthIndicatorResult, HealthIndicatorService } from '@nestjs/terminus'
-import { Cache } from 'cache-manager'
+import type { Cache } from 'cache-manager'
 
 @Injectable()
 export class RedisHealthIndicator {